requirements.txt
numpy==1.26.3
python-dotenv==1.0.0
opencv-python==4.8.1.78
scenedetect==0.6.5.2
torch
torchvision
torchaudio
ultralytics
boto3
mediapipe
hydra-core==1.3.2
pose-format
pytorch-lightning
vector-quantize-pytorch